# openAPI_Spec/admin端接口.openapi.yaml
openapi: 3.1.0
info:
  title: admin端接口
  description: ''
  version: 1.0.0
...
paths:
  /admin/category:
    put:
      summary: 修改分类
      deprecated: false
      description: ''
      tags:
        - 分类相关接口
      parameters:
        - name: Content-Type
          in: header
          description: ''
          required: true
          example: application/json
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  format: int64
                  description: 分类id
                name:
                  type: string
                  description: 分类名称
                sort:
                  type: integer
                  format: int32
                  description: 排序
                type:
                  type: integer
                  format: int32
                  description: 分类类型
              title: CategoryDTO
              $$ref: '#/definitions/CategoryDTO'
              required:
                - id
                - name
                - sort
                - type
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  data:
                    type: string
                  msg:
                    type: string
                title: R«string»
                $$ref: '#/definitions/R«string»'
                required:
                  - code
          headers: {}
      security: []
    post:
      summary: 新增分类
      deprecated: false
      description: ''
      tags:
        - 分类相关接口
      parameters:
        - name: Content-Type
          in: header
          description: ''
          required: true
          example: application/json
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  format: int64
                name:
                  type: string
                  description: 分类名称
                sort:
                  type: integer
                  format: int32
                  description: 排序，按照升序排序
                type:
                  type: integer
                  format: int32
                  description: 分类类型：1为菜品分类，2为套餐分类
              title: CategoryDTO
              $$ref: '#/definitions/CategoryDTO'
              required:
                - name
                - sort
                - type
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  data:
                    type: string
                  msg:
                    type: string
                title: R«string»
                $$ref: '#/definitions/R«string»'
                required:
                  - code
          headers: {}
      security: []
    delete:
      summary: 根据id删除分类
      deprecated: false
      description: ''
      tags:
        - 分类相关接口
      parameters:
        - name: id
          in: query
          description: 分类id
          required: true
          example: '100'
          schema:
            type: string
      requestBody:
        content:
          text/plain:
            schema:
              type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  data:
                    type: string
                  msg:
                    type: string
                title: R«string»
                $$ref: '#/definitions/R«string»'
                required:
                  - code
          headers: {}
      security: []
  /admin/category/page:
    get:
      summary: 分类分页查询
      deprecated: false
      description: ''
      tags:
        - 分类相关接口
      parameters:
        - name: name
          in: query
          description: 分类名称
          required: false
          example: 传统主食
          schema:
            type: string
        - name: page
          in: query
          description: 页码
          required: true
          example: '1'
          schema:
            type: string
        - name: pageSize
          in: query
          description: 每页记录数
          required: true
          example: '10'
          schema:
            type: string
        - name: type
          in: query
          description: 分类类型：1为菜品分类，2为套餐分类
          required: false
          example: '1'
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: number
                  msg:
                    type: 'null'
                  data:
                    type: object
                    properties:
                      total:
                        type: number
                      records:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: number
                            type:
                              type: number
                            name:
                              type: string
                            sort:
                              type: number
                            status:
                              type: number
                            createTime:
                              type: string
                            updateTime:
                              type: string
                            createUser:
                              type: number
                            updateUser:
                              type: number
                          required:
                            - id
                            - type
                            - name
                            - sort
                            - status
                            - createTime
                            - updateTime
                            - createUser
                            - updateUser
                    required:
                      - total
                      - records
                required:
                  - code
                  - data
          headers: {}
      security: []
  /admin/category/status/{status}:
    post:
      summary: 启用、禁用分类
      deprecated: false
      description: ''
      tags:
        - 分类相关接口
      parameters:
        - name: status
          in: path
          description: 1为启用，0为禁用
          required: true
          example: '1'
          schema:
            type: string
        - name: id
          in: query
          description: 分类id
          required: true
          example: '100'
          schema:
            type: string
        - name: Content-Type
          in: header
          description: ''
          required: true
          example: application/json
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: {}
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  data:
                    type: string
                  msg:
                    type: string
                title: R«string»
                $$ref: '#/definitions/R«string»'
                required:
                  - code
          headers: {}
      security: []
  /admin/category/list:
    get:
      summary: 根据类型查询分类
      deprecated: false
      description: ''
      tags:
        - 分类相关接口
      parameters:
        - name: type
          in: query
          description: 分类类型：1为菜品分类，2为套餐分类
          required: false
          example: '2'
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        createTime:
                          type: string
                          format: date-time
                        createUser:
                          type: integer
                          format: int64
                        id:
                          type: integer
                          format: int64
                        name:
                          type: string
                        sort:
                          type: integer
                          format: int32
                        status:
                          type: integer
                          format: int32
                        type:
                          type: integer
                          format: int32
                        updateTime:
                          type: string
                          format: date-time
                        updateUser:
                          type: integer
                          format: int64
                      title: Category
                      $$ref: '#/definitions/Category'
                  msg:
                    type: string
                title: R«List«Category»»
                $$ref: '#/definitions/R«List«Category»»'
                required:
                  - code
          headers: {}
      security: []
...
components:
...
  securitySchemes: {}
servers: []
security: []